# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "D-Care Console API",
    "description": "치과 아웃바운드 관리 시스템 API\n\n## 인증\n모든 V2 API는 JWT Bearer 토큰 인증이 필요합니다.\n`Authorization: Bearer <token>` 헤더를 포함하세요.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "현재 서버" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalCount": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "phone": { "type": "string" },
          "status": { "type": "string", "enum": ["consulting", "reserved", "visited", "treatmentBooked", "treatment", "completed", "followup", "closed"] },
          "temperature": { "type": "string", "enum": ["hot", "warm", "cold"] },
          "interest": { "type": "string" },
          "source": { "type": "string" },
          "consultationType": { "type": "string" },
          "age": { "type": "integer" },
          "urgency": { "type": "string", "enum": ["noshow", "today", "overdue", "normal"] },
          "daysInStatus": { "type": "integer" },
          "nextAction": { "type": "string" },
          "nextActionDate": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastContactAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v2/patients": {
      "get": {
        "tags": ["Patients"],
        "summary": "환자 목록 조회",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["consulting", "reserved", "visited", "treatmentBooked", "treatment", "completed", "followup", "closed"] } },
          { "name": "temperature", "in": "query", "schema": { "type": "string", "enum": ["hot", "warm", "cold"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "이름 또는 전화번호 검색" },
          { "name": "urgency", "in": "query", "schema": { "type": "string", "enum": ["noshow", "today", "overdue", "normal"] } },
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["1month", "3months", "6months", "1year", "all"], "default": "3months" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "createdAt" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } }
        ],
        "responses": {
          "200": {
            "description": "환자 목록",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "patients": { "type": "array", "items": { "$ref": "#/components/schemas/Patient" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" },
                    "filterStats": { "type": "object", "description": "상태별 환자 수" },
                    "urgentStats": { "type": "object", "properties": { "noshow": { "type": "integer" }, "today": { "type": "integer" }, "overdue": { "type": "integer" } } }
                  }
                }
              }
            }
          },
          "401": { "description": "인증 필요" }
        }
      },
      "post": {
        "tags": ["Patients"],
        "summary": "신규 환자 등록",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "phone"],
                "properties": {
                  "name": { "type": "string" },
                  "phone": { "type": "string" },
                  "consultationType": { "type": "string" },
                  "interest": { "type": "string", "description": "관심 시술" },
                  "source": { "type": "string", "description": "유입 경로" },
                  "temperature": { "type": "string", "enum": ["hot", "warm", "cold"], "default": "warm" },
                  "age": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "등록 성공",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "patientId": { "type": "string" } } } } }
          },
          "409": { "description": "전화번호 중복" }
        }
      }
    },
...
  },
...
}